Backend/src/main/java/com/talan/adminmodule/service/impl/CategoryServiceImpl.java
package com.talan.adminmodule.service.impl;
import com.talan.adminmodule.dto.CategoryDto;
import com.talan.adminmodule.entity.Category;
import com.talan.adminmodule.repository.CategoryRepository;
import com.talan.adminmodule.service.CategoryService;
import jakarta.persistence.EntityNotFoundException;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;
import java.util.stream.Collectors;

@Service
public class CategoryServiceImpl implements CategoryService {
    private final CategoryRepository categoryRepository;

    @Autowired
    public CategoryServiceImpl(CategoryRepository categoryRepository) {
        this.categoryRepository = categoryRepository;
    }

    @Override
    public CategoryDto save(CategoryDto categoryDto) {
        return CategoryDto.fromEntity(categoryRepository.save(CategoryDto.toEntity(categoryDto)));
    }

    @Override
    public void delete(Integer id) {
        this.categoryRepository.deleteById(id);
    }

    @Override
    public CategoryDto findById(Integer id) {
        Category category = this.categoryRepository.findById(id).orElseThrow(() -> new EntityNotFoundException(
                "No category with ID = " + id + " found in the database")
        );
        return CategoryDto.fromEntity(category);
    }

    @Override
    public List<CategoryDto> findAll() {
        return categoryRepository.findAll().stream()
                .map(CategoryDto::fromEntity)
                .toList();
    }
}
